ui-optimize：多种布局类型列配置调整优化

+++ ui-ngx/src/app/modules/formly/layout/array/formly-field-array.component.ts
@@ -6,31 +6,41 @@ import {
   CdkDropList,
 } from '@angular/cdk/drag-drop';
 import { Component, OnInit } from '@angular/core';
+import { FormGroup } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIcon } from '@angular/material/icon';
-import { FieldArrayType, FormlyModule } from '@ngx-formly/core';
+import {
+  FieldArrayType,
+  FormlyFormBuilder,
+  FormlyModule,
+} from '@ngx-formly/core';
 import { FormEditorService } from '@src/app/core/services/form-editor.service';
+import { deepClone } from '@src/app/core/utils';
 import { IEditorFormlyField } from '@src/app/shared/models/widget.model';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'formly-field-array',
   templateUrl: './formly-field-array.component.html',
   styleUrls: ['./formly-field-array.component.less'],
   imports: [
+    MatIcon,
     CdkDrag,
     CdkDropList,
-    CdkDragHandle,
-    CdkDragPlaceholder,
     FormlyModule,
-    MatIcon,
+    CdkDragHandle,
     MatButtonModule,
+    CdkDragPlaceholder,
   ],
 })
 export class FormlyFieldArrayComponent
   extends FieldArrayType<IEditorFormlyField>
   implements OnInit
 {
-  constructor(private formEditorService: FormEditorService) {
+  constructor(
+    private toastr: ToastrService,
+    private formEditorService: FormEditorService,
+  ) {
     super();
   }
 
@@ -44,8 +54,13 @@ export class FormlyFieldArrayComponent
   }
 
   addField(toIndex: number) {
+    let defaultAddValue = this.field.props?.['defaultAddValue'];
+    defaultAddValue && (defaultAddValue = deepClone(defaultAddValue));
+
+    const value = deepClone(this.model.at(-1) || {});
+
     this.formEditorService.addField(
-      this.model.at(-1),
+      defaultAddValue || value,
       this.model,
       toIndex,
       false,
@@ -54,6 +69,15 @@ export class FormlyFieldArrayComponent
   }
 
   removeField(toIndex: number) {
+    const canRemoveLast = this.field.props?.['canRemoveLast'] ?? true;
+    // 如果不允许删除最后一列，且当前只有一列，则提示不能删除
+    if (!canRemoveLast && this.model.length === 1) {
+      this.toastr.warning('最后一列不能删除', '', {
+        positionClass: 'toast-top-center',
+      });
+      return;
+    }
+
     this.formEditorService.removeField(this.model, toIndex, false);
     this.options.build!();
   }
@@ -63,10 +87,8 @@ export class FormlyFieldArrayComponent
     itemField: IEditorFormlyField,
   ): string {
     const itemRow = itemField.props?.['row'] ?? 1;
-    const totalRow = fieldGroup.reduce(
-      (acc, ori) => acc + (ori.props?.['row'] ?? 0),
-      0,
-    ) || 1;
+    const totalRow =
+      fieldGroup.reduce((acc, ori) => acc + (ori.props?.['row'] ?? 0), 0) || 1;
     return `calc(100% * ${itemRow / totalRow}) `;
   }
